on_chain/requirements.txt
web3==6.15.1
py-solc-x==2.0.2
python-dotenv==1.0.1 